New projection component for use from any view
